Re-run signature verifier for annotate-exports codemod
Summary:
Before this diff, the annotate-exports codemod would work by reading the file
sig from shared memory (in the typed runner), then extracting the tolerable
errors from the file sig record (in the codemod) to use as the starting point
for the codemod.

I would like to stop storing errors in the file sig, which means that the errors
would no longer be available in the worker. To work around this limitation, this
diff changes the annotate-exports codemod to perform signature verification in
the worker itself.

Signature verification is fast and this codemod does not need to be very fast,
so I am not concerned with performance here. I tried to be unobtrusive with the
changes for ease of reviewing, but welcome suggestions to improve clarity here.

Reviewed By: panagosg7

Differential Revision: D22521930

fbshipit-source-id: c42a5635c85c54141386ce38d232e7801ae47875